requirements.txt
numpy>=1.19.0
qrcode>=7.3.1
Pillow>=9.0.0
opencv-python>=4.5.0